[rename] making authentication module
