+++ src/client/admin/end-user-agreement/admin-eua.module.ts
@@ -10,6 +10,7 @@ import { AdminService } from '../admin.service';
 import { AdminListEuasComponent } from './admin-list-euas.component';
 import { AdminCreateEuaComponent } from './admin-create-eua.component';
 import { AdminUpdateEuaComponent } from './admin-edit-eua.component';
+import { UserEuaComponent } from './user-eua.component';
 import { UtilModule } from '../../shared/util.module';
 import { AlertService } from '../../shared/alert.service';
 
@@ -25,7 +26,8 @@ import { AlertService } from '../../shared/alert.service';
 	declarations:   [
 		AdminListEuasComponent,
 		AdminCreateEuaComponent,
-		AdminUpdateEuaComponent
+		AdminUpdateEuaComponent,
+		UserEuaComponent
 	],
 	providers:  [
 		AlertService,
