+++ src/client/app-routing.module.ts
@@ -1,17 +1,17 @@
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
-import { AuthGuard } from './core/services/auth-guard.service';
-import { SigninComponent } from './admin/components/signin.client.component';
-import { UnauthorizedComponent } from './admin/components/unauthorized.client.component';
-import { UserSignupComponent } from './admin/components/user-management/user-signup.client.component';
-import { UpdateUserComponent } from './admin/components/user-management/edit-user.client.component';
-import { InvalidResourceComponent } from './core/components/invalid-resource.client.component';
-import { InvalidCertificateComponent } from './admin/components/invalid-certificate.client.component';
+import { AuthGuard } from './core/auth-guard.service';
+import { SigninComponent } from './admin/signin.component';
+import { UnauthorizedComponent } from './admin/unauthorized.component';
+import { UserSignupComponent } from './admin/user-management/user-signup.component';
+import { UpdateUserComponent } from './admin/user-management/edit-user.component';
+import { InvalidResourceComponent } from './core/invalid-resource.component';
+import { InvalidCertificateComponent } from './admin/invalid-certificate.component';
 import { AuditComponent } from './audit/components/audit.client.component';
-import { UserEuaComponent } from './admin/components/end-user-agreement/user-eua.client.component';
-import { NoAccessComponent } from './admin/components/no-access.client.component';
-import { LoggedInComponent } from './core/components/logged-in.client.component';
-import { InactiveComponent } from './admin/components/inactive.client.component';
+import { UserEuaComponent } from './admin/end-user-agreement/user-eua.component';
+import { NoAccessComponent } from './admin/no-access.component';
+import { LoggedInComponent } from './core/logged-in.component';
+import { InactiveComponent } from './admin/inactive.component';
 
 @NgModule({
 	imports: [
