+++ src/client/core/core.component.ts
@@ -1,6 +1,3 @@
-import { Component } from '@angular/core';
-import { Router, NavigationEnd } from '@angular/router';
-
 import { ConfigService } from './config.service';
 import { User } from '../admin/user.class';
 import { AuthenticationService } from '../admin/authentication.service';
@@ -32,64 +29,5 @@ export class CoreComponent {
 				this.copyright = config.copyright;
 				this.pki = config.auth === 'proxy-pki';
 			} );
-	}
-}
-
-/**
- * Primary header component wrapping the whole page
- */
-@Component({
-	selector: 'core-header',
-	templateUrl: '../views/header.client.view.html'
-})
-export class HeaderComponent extends CoreComponent {
-	private currentRoute: string = '';
-
-	constructor(
-		protected auth: AuthenticationService,
-		protected configService: ConfigService,
-		protected router: Router
-	) {
-		super(auth, configService);
-	}
-
-	ngOnInit() {
-		super.ngOnInit();
-
-		this.router.events.subscribe( (event) => {
-			if (event instanceof NavigationEnd) {
-				this.currentRoute = event.url;
-			}
-		});
-	}
-}
-
-/**
- * Primary footer component wrapping the whole page
- */
-@Component({
-	selector: 'core-footer',
-	templateUrl: '../views/footer.client.view.html'
-})
-
-export class FooterComponent extends CoreComponent {
-	private currentRoute: string = '';
-
-	constructor(
-		protected auth: AuthenticationService,
-		protected configService: ConfigService,
-		protected router: Router
-	) {
-		super(auth, configService);
-	}
-
-	ngOnInit() {
-		super.ngOnInit();
-
-		this.router.events.subscribe( (event) => {
-			if (event instanceof NavigationEnd) {
-				this.currentRoute = event.url;
-			}
-		});
 	}
 }
