+++ src/client/core/core.scss
@@ -1,4 +1,4 @@
-@import '../../shared/sass/colors';
+@import '../shared/colors';
 
 //=====================================
 // Fix for Modal backdrop
