+++ src/client/help/help.scss
@@ -1,4 +1,4 @@
-@import '../../shared/sass/colors';
+@import '../shared/colors';
 
 .help-content {
 	padding-top: 25px;
