+++ src/client/img/brand/logo-dark-bg.png
