+++ src/client/img/brand/logo-light-bg.png
