+++ src/client/shared/general.scss
@@ -1,4 +1,4 @@
-@import '../../shared/sass/colors';
+@import './colors';
 
 //==================================
 // General Purpose Styles
